Switch to date-based log rotation
Considering these files can be quite large, and we are very much
interested in archiving them, make logrotate append a date-stamp
rather than a number.

This makes them work better with archival tools (such as duplicity),
and makes for less head scratching on later analysis.

In the process, make the duration we keep the logs on disk locally
configurable. 365 days is nice (and the previous default), but that
could chew up a lot of space.