Fix ngnix worker log rotation - using wrong user
We use the 'nginx' user for nginx, rather than the package default of
'www-data'. This is likely a hangover from a migration from CentOS, but
now we're bedded in we may as well stick with it.

Ensure the log dir is readable/writable by nginx user. Previously it wasn't,
so workers were struggling to reread log files.