opgops-938 - nginx installations fail
Currently use of this formula is futile as the version currently specified
no longer exists in the Ubuntu repository.

The specific version has been removed and an empty string put in its place,
which should allow the package manager to select the latest version.

If a specific version should be installed, this should be specified in the
pillar file.