Check the nginx config explicitly
Nginx will check it's configuration files before attempting to load them.
If the check fails then nginx carries on with the old config without
terminating. Unfortunately this means that the formula user can be unaware that
the configuration changes they have made are not being used.

This change adds a state that always runs a check on nginx configuration
and returns an error if the config is bad.