Unbreak nginx test on initial highstate
Currently there doesn't seem to be a way to lazy evaluate watching
for resources that may get deleted. Previously the initial highstate
breaks because as high state (2nd pass) is running the directory gets
deleted, but the `test -d` is executed during salt's low state (1st
pass). Since we are removing the directory there may be little point
in watching it.